+++ src/pages/welcome-page/components/infoAboutProject/index.tsx
@@ -1,1 +1,1 @@
-export { InfoAboutProject } from './InfoAboutProject';
+export { InfoAboutProject } from './info-about-project';
