+++ src/pages/welcome-page/components/infoAboutProject/info-about-project.css
