+++ src/pages/welcome-page/index.ts
@@ -1,1 +1,1 @@
-export { WelcomePage } from './WelcomePage';
+export { WelcomePage } from './welcome-page';
