+++ src/pages/welcome-page/welcome-page.tsx
@@ -1,4 +1,4 @@
-import './WelcomePage.css';
+import './welcome-page.css';
 import { InfoAboutProject } from './components/infoAboutProject/index';
 
 export const WelcomePage = () => (
